Desktop/stock-dashboard/requirements.txt
altair==5.5.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
captcha==0.7.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
curl_cffi==0.11.4
extra-streamlit-components==0.1.80
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
multitasking==0.0.11
narwhals==1.44.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.46.1
streamlit-authenticator==0.4.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.64
